+++ src/tests/test07_Subscription/main.cpp
@@ -117,7 +117,10 @@ void setup() {
 
     // 默认目标速度, 避免订阅消息到达前电机无目标
     // 逆解输出仅本次使用, 声明为局部变量, 作用域最小化 (仿照 main.cpp)
-    const float body_velocities[2] = {DEFAULT_LINEAR_SPEED_MM_S, DEFAULT_ANGULAR_SPEED_RAD_S}; // 车体速度: [0]=线速度 mm/s, [1]=角速度 rad/s
+    const float body_velocities[2] = {
+        DEFAULT_LINEAR_SPEED_MM_S,
+        DEFAULT_ANGULAR_SPEED_RAD_S
+    };                     // 车体速度: [0]=线速度 mm/s, [1]=角速度 rad/s
     float motor_speeds[2]; // 电机转速: [0]=左, [1]=右, 单位 mm/s, 仅用于本次逆解计算
     kinematics.kinematics_inverse(body_velocities, motor_speeds);
 
@@ -150,8 +153,8 @@ void twist_callback(const void* msg_in) {
     // 运动学逆解: 车体速度 -> 电机目标转速
     // linear.x 单位 m/s 转换为 mm/s, angular.z 单位 rad/s
     const float body_velocities[2] = {
-        twist_msg->linear.x * MPS_TO_MMPS, // [0]=线速度, 单位换算 m/s -> mm/s
-        twist_msg->angular.z               // [1]=角速度, 单位 rad/s
+        static_cast<float>(twist_msg->linear.x * MPS_TO_MMPS), // [0]=线速度, 单位换算 m/s -> mm/s
+        static_cast<float>(twist_msg->angular.z)               // [1]=角速度, 单位 rad/s
     };
     float motor_speeds[2]; // 电机转速: [0]=左, [1]=右, 单位 mm/s, 仅用于本次逆解计算
     kinematics.kinematics_inverse(body_velocities, motor_speeds);
@@ -218,7 +221,10 @@ void micro_ros_task(void* parameter) {
  * 再经 PID 控制器输出 PWM 值更新电机。
  */
 void update_and_control() {
-    const int32_t ticks[2] = {encoders[0].getTicks(), encoders[1].getTicks()}; // 编码器 tick: [0]=左, [1]=右
+    const int32_t ticks[2] = {
+        encoders[0].getTicks(),
+        encoders[1].getTicks()
+    }; // 编码器 tick: [0]=左, [1]=右
     kinematics.update_motor_speed(millis(), ticks);
 
     motor.updateMotorSpeed(0, pid_controller[0].update_pwm(kinematics.get_motor_speed(0)));
